requirements.txt
asyncpg
click
git+https://github.com/Rapptz/discord.py@rewrite
pillow
pyfiglet
ruamel.yaml
unidecode
uvloop
